+++ api_service/app/internal/transport/http/v1/pro/putgm_image.go
@@ -3,6 +3,8 @@ package pro
 import (
 	"fmt"
 	"net/http"
+	"os"
+	"path/filepath"
 
 	"github.com/Alexander272/sealur/api_service/internal/models"
 	"github.com/Alexander272/sealur/api_service/internal/transport/http/v1/proto"
@@ -53,6 +55,7 @@ func (h *Handler) getPutgmImage(c *gin.Context) {
 // @Security ApiKeyAuth
 // @Description создание чертежа путгм
 // @ModuleID createPutgmImage
+// @Accept multipart/form-data
 // @Accept json
 // @Produce json
 // @Param data body models.PutgmImageDTO true "putgm image info"
@@ -62,11 +65,29 @@ func (h *Handler) getPutgmImage(c *gin.Context) {
 // @Failure default {object} models.ErrorResponse
 // @Router /sealur-pro/putgm-image [post]
 func (h *Handler) createPutgmImage(c *gin.Context) {
-	//TODO дописать добавление картинки
-
 	var dto models.PutgmImageDTO
-	if err := c.BindJSON(&dto); err != nil {
-		models.NewErrorResponse(c, http.StatusBadRequest, err.Error(), "invalid data send")
+
+	if c.Request.FormValue("gasket") != "" {
+		dto.Gasket = c.Request.FormValue("gasket")
+		dto.Form = c.Request.FormValue("form")
+		dto.Url = c.Request.FormValue("url")
+	} else {
+		if err := c.BindJSON(&dto); err != nil {
+			models.NewErrorResponse(c, http.StatusBadRequest, err.Error(), "invalid data send")
+			return
+		}
+	}
+
+	var folder string
+	if dto.Form == "Round" {
+		folder = "putgm/construction"
+	} else {
+		folder = "putgm/half"
+	}
+
+	err := h.saveFile(c, folder)
+	if err != nil {
+		models.NewErrorResponse(c, http.StatusBadRequest, err.Error(), "error while saving file")
 		return
 	}
 
@@ -84,6 +105,19 @@ func (h *Handler) createPutgmImage(c *gin.Context) {
 
 	c.Header("Location", fmt.Sprintf("/api/v1/sealur-pro/putgm-image/%s", image.Id))
 	c.JSON(http.StatusCreated, models.IdResponse{Id: image.Id, Message: "Created"})
+}
+
+func (h *Handler) saveFile(c *gin.Context, folder string) error {
+	file, err := c.FormFile("drawing")
+	if err != nil {
+		return err
+	}
+
+	if err := c.SaveUploadedFile(file, filepath.Join("images", folder, file.Filename)); err != nil {
+		return err
+	}
+
+	return nil
 }
 
 // @Summary Update Putgm Image
@@ -137,18 +171,25 @@ func (h *Handler) updatePutgmImage(c *gin.Context) {
 // @Accept json
 // @Produce json
 // @Param id path string true "putgm image id"
-// @Success 201 {object} models.IdResponse
+// @Param file query string true "file name"
+// @Success 200 {object} models.IdResponse
 // @Failure 400,404 {object} models.ErrorResponse
 // @Failure 500 {object} models.ErrorResponse
 // @Failure default {object} models.ErrorResponse
 // @Router /sealur-pro/putgm-image/{id} [delete]
 func (h *Handler) deletePutgmImage(c *gin.Context) {
-	//TODO дописать удаление картинки
 	id := c.Param("id")
 	if id == "" {
 		models.NewErrorResponse(c, http.StatusBadRequest, "empty id", "empty id param")
 		return
 	}
+
+	file := c.Query("file")
+	if file == "" {
+		models.NewErrorResponse(c, http.StatusBadRequest, "empty file", "empty file param")
+		return
+	}
+	os.Remove(filepath.Join("images", file))
 
 	image, err := h.proClient.DeletePutgmImage(c, &proto.DeletePutgmImageRequest{Id: id})
 	if err != nil {
